EDK: Change Awakable::Awake()'s context, etc., to be a uint64_t.
This is in preparation for implementing the wait set API that was added
in 91a7a240f900.

R=azani@chromium.org
B=#350

Review URL: https://codereview.chromium.org/2056763003 .

+++ mojo/edk/system/core.h
@@ -230,7 +230,7 @@ class Core {
                               const MojoHandleSignals* signals,
                               uint32_t num_handles,
                               MojoDeadline deadline,
-                              uint32_t* result_index,
+                              uint64_t* result_index,
                               HandleSignalsState* signals_states);
 
   embedder::PlatformSupport* const platform_support_;
