+++ mojo/edk/system/simple_dispatcher.h
@@ -32,7 +32,7 @@ class SimpleDispatcher : public Dispatcher {
   void CancelAllStateNoLock() override;
   MojoResult AddAwakableImplNoLock(Awakable* awakable,
                                    MojoHandleSignals signals,
-                                   uint32_t context,
+                                   uint64_t context,
                                    HandleSignalsState* signals_state) override;
   void RemoveAwakableImplNoLock(Awakable* awakable,
                                 HandleSignalsState* signals_state) override;
